feat(stock): support fractional quantities (foundation for raw materials)
Stock and stock-history quantities become NUMERIC(14,3) so materials measured
in mm/m/kg/L can be tracked accurately (e.g. 2.5 m). Server validation drops
.int() on quantity/delta, valuation SUM no longer truncates via ::int, and the
location editor parses decimals. OpenAPI spec updated (orval output unchanged —
TS maps integer and number identically).

Migration: 0024_fractional_stock.sql

+++ artifacts/api-server/src/routes/stock.ts
@@ -14,7 +14,7 @@ router.get("/valuation", async (req, res) => {
     const totals = await db
       .select({
         productId: stockTable.productId,
-        totalQty: sql<number>`COALESCE(SUM(${stockTable.quantity}), 0)::int`.as("total_qty"),
+        totalQty: sql<number>`COALESCE(SUM(${stockTable.quantity}), 0)::numeric`.as("total_qty"),
       })
       .from(stockTable)
       .innerJoin(locationsTable, eq(stockTable.locationId, locationsTable.id))
@@ -108,8 +108,8 @@ router.get("/:locationId", async (req, res) => {
 });
 
 const updateStockSchema = z.object({
-  quantity: z.number().int().min(0).optional(),
-  delta: z.number().int().optional(),
+  quantity: z.number().min(0).optional(),
+  delta: z.number().optional(),
   changedBy: z.string().nullable().optional(),
 });
